% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_GMT_time}
\alias{add_GMT_time}
\title{Add a column of GMT time to a cyan data query}
\usage{
add_GMT_time(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with an additional character column for time in GMT
}
\description{
Find GMT time given the sample time and local time zone
}
