% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_solar_noon}
\alias{add_solar_noon}
\title{Add a logical column indicating whether the sample was taken during solar noon}
\usage{
add_solar_noon(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with two additional logical columns, solar_noon indicating
if the sample was taken during solar noon and ext_solar_noon indicating if
the sample was taken during extended solar noon.
}
\description{
Solar noon is defined as 1000 - 1400, extended solar noon is defined as
0900 - 1500
}
