% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_trophic_status}
\alias{add_trophic_status}
\title{Add columns for trophic status}
\usage{
add_trophic_status(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with three additional columns. TROPHIC_STATUS_INDEX
 giving the numerical trophic status index,
 TROPHIC_STATUS - indicating which
trophic category that index falls into, and the method used to
calculate the trophic stats. TROPHIC_STATUS_METHOD
}
\description{
Add three columns for trophic status, where it is possible. Trophic
status can be calculated from secchi depth, chlorophyll-a, or
total phosphorus
}
\examples{

#Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)

#Get all of the chlorophyll and chlorophyll-a data (parameter id P0051 & P0054)
#or the state of Kansas in the year 2016
ks_chl_2016 <- get_cyan_data(db_connection, collect = TRUE,
                             years = 2016,
                             parameters = c("P0051", "P0054"),
                             states = "KS")

#Add trophic status columns
ks_chl_2016_wtrophic <- add_trophic_status(ks_chl_2016)

}
