% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_WHO_category}
\alias{add_WHO_category}
\title{Add WHO thresholds}
\usage{
add_WHO_category(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with an additional column WHO_CATEGORY indicating whether
this result is classified as Low, Moderate, High, or Very High based on
WHO thresholds for the parameter
}
\description{
Add columns indicating the WHO category for Cyanobacteria, Microcystin-LR, or
Cholorophyll-a, if applicable
}
