% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{apply_flags}
\alias{apply_flags}
\title{Apply a given flag to results}
\usage{
apply_flags(cyan_connection, flag_code, initials, results)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{flag_code}{the flag code to apply}

\item{initials}{initials to tag the flag with}

\item{results}{the results to apply the flag to}
}
\value{
a logical indicating whether the flags were succesfully written
}
\description{
Add flags to the QCFLAGS table as defined in the FLAG_KEY table to a given
vector of RESULT_IDs
}
