% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_cyan}
\alias{connect_cyan}
\title{Connect to the CyAN database}
\usage{
connect_cyan(path)
}
\arguments{
\item{path}{a character string giving the path to the database}
}
\value{
a SQLite database connection
}
\description{
Generate a connection to the CyAN database or another sqlite database using
the CyAN schema and parameter table
}
\examples{
#Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)

}
