% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyan_reditools.R
\name{.cyan_reditools_build_args}
\alias{.cyan_reditools_build_args}
\title{Build the REDItools2 CLI argument vector}
\usage{
.cyan_reditools_build_args(
  bam,
  reference,
  output_file,
  strand_mode,
  min_coverage,
  min_base_quality,
  min_read_quality,
  min_read_length,
  min_edits,
  strict,
  region,
  omopolymeric_file,
  splicing_file,
  bed_file,
  debug
)
}
\description{
Build the REDItools2 CLI argument vector
}
\keyword{internal}
