% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagging.R
\name{find_flagged_data}
\alias{find_flagged_data}
\title{Find results with a particular flag}
\usage{
find_flagged_data(cyan_connection, flag_code, collect = TRUE)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{flag_code}{the flag code of interest}

\item{collect}{a logical indicating whether the query will be pulled into
a local tibble using dbplyr::collect. If you are planning on doing further
selection or filtering, you may want not want to collect until you're ready}
}
\value{
a table of all results that have been flagged with the specified flag code
}
\description{
Check the QC_FLAGS table to find results that have been flagged with
the given flag code
}
