% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_tools.R
\name{null_if_blank_as_num}
\alias{null_if_blank_as_num}
\title{Return a null value for an empty character string}
\usage{
null_if_blank_as_num(x)
}
\arguments{
\item{x}{a blank character string or number as text}
}
\value{
either NULL or a number
}
\description{
If x is a blank character vector, return null, otherwise make sure x is numeric
}
