% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{Plot bivariate data}
\usage{
plot_bivariate(plot_data)
}
\arguments{
\item{plot_data}{bivariate data to plot - should be collected
before the plotting function}

\item{method_highlight}{a method to be highlighted - points analyzed
with this method will appear in red instead of black}

\item{flagged_results}{a vector of flagged results - could be
obtained using the \code{find_flagged} function}
}
\description{
Plot data retrieved using \code{get_bivariate}
}
