% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyan_reditools.R
\name{print.cyan_batch_result}
\alias{print.cyan_batch_result}
\title{Print method for cyan_batch_result}
\usage{
\method{print}{cyan_batch_result}(x, ...)
}
\description{
Print method for cyan_batch_result
}
