% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{remove_flags}
\alias{remove_flags}
\title{Remove flags from the QC_FLAGS table}
\usage{
remove_flags(cyan_connection, flag_code, results)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{flag_code}{the flag code to remove}

\item{results}{the result identifiers to remove the given flag from}
}
\value{
a numeric vector indicating the number of flags deleted for each result
}
\description{
Given a flag_code and a vector of result identifiers, remove all the applicable
rows from the QC_FLAGS table
}
