% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_EPA_recreational_threshold}
\alias{add_EPA_recreational_threshold}
\title{Add EPA recreational threshold}
\usage{
add_EPA_recreational_threshold(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with an additional column EPA_REC_EXCEEDED - a logical column
indicating whether the applicable EPA recreational treshold was exceeded
}
\description{
Add a column indicating whether the EPA recreational trehsold was exceeded
}
