% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_GMT_time}
\alias{add_GMT_time}
\title{Add a column of GMT time to a cyan data query}
\usage{
add_GMT_time(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data()} with
\code{collect = TRUE}}
}
\value{
cyan_data with an additional character column for time in GMT
}
\description{
Give a column of GMT time based on the activity start time and local time zone
}
\examples{
#Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)

#Get all of the chlorophyll and chlorophyll-a data (parameter id P0051 & P0054)
#or the state of Kansas in the year 2016
ks_chl_2016 <- get_cyan_data(db_connection, collect = TRUE,
                             years = 2016,
                             parameters = c("P0051", "P0054"),
                             states = "KS")

#Add the column for GMT time
ks_chl_2016_wGMT <- add_GMT_time(ks_chl_2016)


}
