% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_solar_noon}
\alias{add_solar_noon}
\title{Add a logical column indicating whether the sample was taken during solar noon}
\usage{
add_solar_noon(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data()} with
\code{collect = TRUE}}
}
\value{
cyan_data with two additional logical columns, solar_noon indicating
if the sample was taken during solar noon and ext_solar_noon indicating if
the sample was taken during extended solar noon.
}
\description{
Solar noon is defined as 1000 - 1400, extended solar noon is defined as
0900 - 1500
}
\examples{

#Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)

#Get all of the chlorophyll and chlorophyll-a data (parameter id P0051 & P0054)
#or the state of Kansas in the year 2016
ks_chl_2016 <- get_cyan_data(db_connection, collect = TRUE,
                             years = 2016,
                             parameters = c("P0051", "P0054"),
                             states = "KS",)

#Add logical columns for solar noon and extended solar noon
ks_chl_2016_wsolarnoon <- add_solar_noon(ks_chl_2016)

}
