% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_trophic_status}
\alias{add_trophic_status}
\title{Add columns for trophic status}
\usage{
add_trophic_status(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with three additional columns. TROPHIC_STATUS_INDEX
 giving the numerical trophic status index,
 TROPHIC_STATUS - indicating which
trophic category that index falls into, and the method used to
calculate the trophic stats. TROPHIC_STATUS_METHOD
}
\description{
Add three columns for trophic status, where it is possible. Trophic
status can be calculated from secchi depth, chlorophyll-a, or
total phosphorus
}
