% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_WHO_category}
\alias{add_WHO_category}
\title{Add WHO thresholds}
\usage{
add_WHO_category(cyan_data)
}
\arguments{
\item{cyan_data}{a data frame from \code{get_cyan_data} with
\code{collect = TRUE}}
}
\value{
cyan_data with an additional column WHO_CATEGORY indicating whether
this result is classified as Low, Moderate, High, or Very High based on
WHO thresholds for the parameter
}
\description{
Add columns indicating the WHO category for Cyanobacteria, Microcystin-LR, or
Cholorophyll-a, if applicable
}
\examples{

#' #Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)

#Get all of the chlorophyll and chlorophyll-a data (parameter id P0051 & P0054)
#or the state of Kansas in the year 2016
ks_chl_2016 <- get_cyan_data(db_connection, collect = TRUE,
                             years = 2016,
                             parameters = c("P0051", "P0054"),
                             states = "KS")

#Add any applicable WHO categories
ks_chl_2016_wWHO <- add_WHO_category(ks_chl_2016)

}
