% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_cyan}
\alias{connect_cyan}
\title{Connect to the CyAN database}
\usage{
connect_cyan(path)
}
\arguments{
\item{path}{the full path to the database}
}
\value{
a SQLite database connection
}
\description{
Given a file path, connect to the CyAN database. The database can be
downloaded manually from the FTP site, or using the download function
}
