% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{.cyan_reditools_env}
\alias{.cyan_reditools_env}
\title{REDItools2 basilisk environment}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
.cyan_reditools_env
}
\description{
Python dependencies for \code{CYAN_run_reditools()}.
Packages pinned to validated, compatible versions.
}
\keyword{internal}
