% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{find_flagged}
\alias{find_flagged}
\title{Find results with a particular flag}
\usage{
find_flagged(cyan_connection, flag_code)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{flag_code}{the flag code of interest}
}
\value{
a vector of result_ids with the given flag
}
\description{
Check the QC_FLAGS table to find results that have been flagged with
the given flag code
}
