% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_location_index}
\alias{generate_location_index}
\title{Generate an index of locations and parameters from the database}
\usage{
generate_location_index(cyan_connection)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}
}
\value{
a data frame of all unique combinations of location and parameter
}
\description{
Based on the CyAN database connection, generate an index of all unique combinations
of locations and parameters that have been measured there
}
