% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_parameter_index}
\alias{generate_parameter_index}
\title{Generate a list of parameter identifiers and their short names}
\usage{
generate_parameter_index(cyan_connection)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}
}
\value{
a data frame of parameter identifiers and their associated
short names
}
\description{
Lookup table relating each arbitrary parameter identifier to the short
version of the parameter name
}
