% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{generate_parameter_index}
\alias{generate_parameter_index}
\title{Generate a table of parameter identifiers and their short names}
\usage{
generate_parameter_index(cyan_connection, has_data = FALSE)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan()}}

\item{has_data}{a logical indicating whether to limit the parameter list to only
those parameters that have data populated in the database}
}
\value{
a data frame of parameter identifiers and their associated
short names
}
\description{
Generate a table of parameter identifiers and their short names
}
\examples{
#Connect to the example database bundled with CyAN
path <- system.file("extdata", "example.db", package = "CyAN")
db_connection <- connect_cyan(path)
#Generate the parameter index
parameter_index <- generate_parameter_index(db_connection)
#Limit the index to the parameters that have data
parameter_index_hasdata <- generate_parameter_index(db_connection, has_data = TRUE)

}
