% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_bivariate}
\alias{get_bivariate}
\title{Find simaltaneous measurements of two parameters}
\usage{
get_bivariate(cyan_connection, parameter_1, parameter_2, collect = FALSE,
  north_latitude = NULL, south_latitude = NULL,
  west_longitude = NULL, east_longitude = NULL, years = NULL)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{parameter_1, parameter_2}{CyAN parameter codes for the two parameters
of interest}

\item{collect}{a logical indicating whether the query will be pulled into
a local tibble using dbplyr::collect. If you are planning on doing further
selection or filtering, you may want not want to collect until you're ready}

\item{north_latitude, south_latitude}{numbers indicating the northern most
and sothern most latitude that will be included in the query. Latitude
should be given as a positive number of degrees north of the equator.}

\item{west_longitude, east_longitude}{nunbers indicating the western most
and eastern most latitude that will be included in the query. Longitude
should be given as a negative number of decimal degrees west of the prime
meridian.}

\item{years}{numeric vector of years that will be included in the query}
}
\value{
if collect is FALSE, the query will be generated, but not collected.
See the documentation on \code{collect} for details. Otherwise, if collect
is TRUE, the query will be pulled into memory and returned as a tibble.
Returns a wide data frame, with a shared activity, depth, and depth_unit
columns, and all other columns suffixed with .1 or .2 for the two parameters.
}
\description{
Find sampling activities that have associated measurements of two parameters
of interest - helpful for bivariate plotting/ investigating the relationship
between two parameters.
}
