% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_cyan_data}
\alias{get_cyan_data}
\title{Get data from CyAN}
\usage{
get_cyan_data(cyan_connection, collect = FALSE, north_latitude = NULL,
  south_latitude = NULL, west_longitude = NULL,
  east_longitude = NULL, start_date = NULL, end_date = NULL,
  parameters = NULL, minimum_tier = NULL, states = NULL)
}
\arguments{
\item{cyan_connection}{a CyAN database connection from \code{connect_cyan}}

\item{collect}{a logical indicating whether the query will be pulled into
a local tibble using dbplyr::collect. If you are planning on doing further
selection or filtering, you may want not want to collect until you're ready}

\item{north_latitude, south_latitude}{numbers indicating the northern most
and sothern most latitude that will be included in the query. Latitude
should be given as a positive number of degrees north of the equator.}

\item{west_longitude, east_longitude}{nunbers indicating the western most
and eastern most latitude that will be included in the query. Longitude
should be given as a negative number of decimal degrees west of the prime
meridian.}

\item{start_date, end_date}{dates can be given as character strings in the
form "yyyy-mm-dd" or as Date objects}

\item{parameters}{a character vector of parameter names that will be
returned in the query}

\item{minimum_tier}{a number between 1.0 and 4.0 indicating the minimum
data tier that will be returned in the query}

\item{states}{a character vector of the two-character state abbreviations
the query will included}
}
\value{
if collect is FALSE, the query will be generated, but not collected.
See the documentation on \code{collect} for details. Otherwise, if collect
is TRUE, the query will be pulled into memory and returned as a tibble. If
your query is complicated, or will return a lot of rows, it could take a
long time - minutes.
}
\description{
Basic data query from the CyAN database based on a number of parameters,
used to drive the data download from the map screen of the shiny app. Can also
be used with the plotting functions
}
