% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{Plot bivariate data}
\usage{
plot_bivariate(plot_data, log_1 = FALSE, log_2 = FALSE,
  method_highlight = NULL, flagged_results = NULL, range_1 = NULL,
  range_2 = NULL)
}
\arguments{
\item{plot_data}{bivariate data to plot - should be collected
before the plotting function}

\item{log_1, log_2}{logicals indicating whether to use a log
scale for parameter 1 and parameter 2, respectively}

\item{method_highlight}{a method to be highlighted - points analyzed
with this method will appear in red instead of black}

\item{flagged_results}{a vector of flagged results - could be
obtained using the \code{find_flagged} function. These results will
be plotted with an x symbol rather than a filled circle}

\item{range_1, range_2}{numeric vectors specifiying the range for
values in parameter 1 and parameter 2, respectively. The minimum
should be given first, followed by the maximum.}
}
\description{
Plot data retrieved using \code{get_bivariate}
}
